Avoid int overflow after year 2038
See https://en.wikipedia.org/wiki/Year_2038_problem

We consistently use long long (64 bit) types for input and output of time_t
except where we get 32-bit int - there we cast it as unsigned int
to allow it to work until year 2106 and assume that nobody needs dates before 1970.
It might also cover for ippAddInteger(..., "time-at-creation", time(NULL));
that I left as is.

This patch was done while reviewing potential year-2038 issues in openSUSE.

note: this is probably incomplete, e.g. there are several checks for INT_MAX.

+++ cups/ppd-util.c
@@ -302,7 +302,7 @@ cupsGetPPD3(http_t     *http,		/* I  - HTTP connection or @code CUPS_HTTP_DEFAUL
       }
       else
       {
-        DEBUG_printf("2cupsGetPPD3: Returning ok, filename=\"%s\", modtime=%ld.", buffer, (long)ppdinfo.st_mtime);
+        DEBUG_printf("2cupsGetPPD3: Returning ok, filename=\"%s\", modtime=%lld.", buffer, (long long)ppdinfo.st_mtime);
         *modtime = ppdinfo.st_mtime;
 	return (HTTP_STATUS_OK);
       }
